+++ .github/workflows/main.yml
@@ -37,8 +37,8 @@ jobs:
         id: evaluator
         uses: betrybe/jest-evaluator-action@v7
         with:
-          repository-test-name: 'betrybe/sd-0x-block12-project-movie-cards-library-tests'
-          repository-test-branch: 'master'
+          repository-test-name: 'betrybe/sd-0x-project-movie-cards-library-tests'
+          repository-test-branch: 'main'
       - name: Store evaluation step
         uses: betrybe/store-evaluation-action@v2
         with:
