+++ src/tests/app.test.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+
+import App from '../App';
+
+describe('5 - Crie um componente `<App />`', () => {
+  let wrapper;
+  it('Renderize `<Header />` dentro do componente `<App />`', () => {
+    wrapper = shallow(<App />);
+    expect(wrapper.find('Header').length).toEqual(1);
+  });
+
+  it('Renderize `<MovieList />` dentro do componente `<App />`', () => {
+    expect(wrapper.find('MovieList').length).toEqual(1);
+  });
+});
