+++ src/tests/movieList.test.js
@@ -1,0 +1,51 @@
+import React from 'react';
+import { mount, shallow } from 'enzyme';
+import MovieCard from '../components/MovieCard';
+import MovieList from '../components/MovieList';
+
+const rating = 4.5;
+const ratingThree = 3;
+const movies = [
+  {
+    title: 'Movie Title 1',
+    subtitle: 'Movie Subtitle 1',
+    storyline: 'Movie Storyline 1',
+    rating,
+    imagePath: 'images/movie_1',
+  },
+  {
+    title: 'Movie Title 2',
+    subtitle: 'Movie Subtitle 2',
+    storyline: 'Movie Storyline 2',
+    rating,
+    imagePath: 'images/movie_2',
+  },
+  {
+    title: 'Movie Title 3',
+    subtitle: 'Movie Subtitle 3',
+    storyline: 'Movie Storyline 3',
+    ratingThree,
+    imagePath: 'images/movie_3',
+  },
+];
+
+describe('2 - Crie um componente `<MovieList />`', () => {
+  let wrapper;
+
+  it('Renderize o componente `<MovieList />`', () => {
+    shallow(<MovieList movies={ movies } />);
+  });
+
+  it('Renderize componentes `<MovieCard />` dentro de `MovieList`', () => {
+    wrapper = shallow(<MovieList movies={ movies } />);
+    expect(wrapper.find(MovieCard).length).toEqual(ratingThree);
+  });
+
+  it('Passe uma key para cada `<MovieCard />` renderizado', () => {
+    wrapper = mount(<MovieList movies={ movies } />);
+    const movieCards = wrapper.find(MovieCard);
+    movieCards.forEach((movieCard, index) => {
+      expect(movieCard.key()).toEqual(movies[index].title);
+    });
+  });
+});
